+++ gateway-api-integrator/tests/conftest.py
@@ -10,5 +10,7 @@ def pytest_addoption(parser):
     Args:
         parser: Pytest parser.
     """
+    parser.addoption("--model", action="store", default=None)
     parser.addoption("--charm-file", action="append")
     parser.addoption("--kube-config", action="store", default="~/.kube/config")
+    parser.addoption("--keep-models", action="store_true", default=False)